LayoutTests/editing/input/ios/update-writing-direction-after-inserting-newline.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<style>
body, html {
    margin: 0;
    width: 100%;
    height: 100%;
    font-family: system-ui;
}

.editor {
    width: 300px;
    height: 300px;
    border: 1px solid lightgray;
}
</style>
</head>
<body>
    <div contenteditable class="editor">
        <p>Hello world</p>
    </div>
</body>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("Verifies that switching to a RTL input mode and then pressing Return automatically sets the base writing direction to RTL");

    const editor = document.querySelector(".editor");

    async function doAfterInputEvent(callback) {
        await UIHelper.callFunctionAndWaitForEvent(callback, editor, "input");
    }

    await UIHelper.activateElementAndWaitForInputSession(editor);
    await UIHelper.setKeyboardInputModeIdentifier("ar");
    await doAfterInputEvent(async () => await UIHelper.enterText("\n"));
    await UIHelper.ensurePresentationUpdate();
    caretRectUsingArabic = await UIHelper.getUICaretViewRect();

    await UIHelper.setKeyboardInputModeIdentifier("en_US");
    await doAfterInputEvent(async () => await UIHelper.keyDown("delete"));
    await UIHelper.ensurePresentationUpdate();

    await doAfterInputEvent(async () => await UIHelper.enterText("\n"));
    await UIHelper.ensurePresentationUpdate();
    caretRectUsingEnglish = await UIHelper.getUICaretViewRect();

    shouldBe("caretRectUsingArabic.top", "caretRectUsingEnglish.top");
    shouldBe("caretRectUsingArabic.width", "caretRectUsingEnglish.width");
    shouldBe("caretRectUsingArabic.height", "caretRectUsingEnglish.height");
    shouldBeGreaterThan("caretRectUsingArabic.left", "caretRectUsingEnglish.left");

    editor.blur();
    await UIHelper.waitForKeyboardToHide();

    finishJSTest();
});
</script>
</html>
